% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{add_footer}
\alias{add_footer}
\title{Add the footer to an existing application.}
\usage{
add_footer(location)
}
\arguments{
\item{location}{path of the existing application.}
}
\description{
Add the footer to an existing application.
}
