% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{add_ui_body}
\alias{add_ui_body}
\title{add_ui_body}
\usage{
add_ui_body(body_elements = NULL, append = FALSE)
}
\arguments{
\item{body_elements}{- List of ui elements to be displayed in application body}

\item{append}{- Add elements to current body elements or remove previous body elements (default = FALSE)}
}
\description{
Builds application body with given configurations and elements. It is called within "ui_body.R".
Check example application for detailed example
}
\section{Shiny Usage}{

Call this function from \code{program/ui_body.R} to set body parameters
}

