% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{add_ui_footer}
\alias{add_ui_footer}
\title{add_ui_footer}
\usage{
add_ui_footer(left = NULL, right = NULL, fixed = FALSE)
}
\arguments{
\item{left}{- Left text}

\item{right}{- Right text}

\item{fixed}{- Whether to fix the navbar to the top. FALSE by default.}
}
\description{
Builds application footer with given configurations and elements. It is called within "ui_footer.R".
Check example application for detailed example
}
\section{Shiny Usage}{

Call this function from \code{program/ui_footer.R} to set footer parameters
}

\examples{
  library(shiny)
  library(bs4Dash)

  # Inside ui_footer.R
  # Left text
  left <- a(href   = "https://periscopeapps.org/",
            target = "_blank",
            "periscope2")
  # Right text
  right <- "2022"

  # -- Register Elements in the ORDER SHOWN in the UI
  add_ui_footer(left, right)

}
