% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{add_ui_header}
\alias{add_ui_header}
\title{add_ui_header}
\usage{
add_ui_header(
  skin,
  status,
  border,
  compact,
  sidebarIcon,
  controlbarIcon,
  fixed,
  left_ui,
  right_ui
)
}
\arguments{
\item{skin}{Sidebar skin. "dark" or "light".}

\item{status}{Sidebar status. Check \code{?bs4Dash::bs4DashNavbar()} for list of valid values}

\item{border}{Whether to separate the navbar and body by a border. TRUE by default.}

\item{compact}{Whether items should be compacted. FALSE by default.}

\item{sidebarIcon}{Icon of the main sidebar toggle}

\item{controlbarIcon}{Icon to toggle the controlbar (left)}

\item{fixed}{Whether to fix the navbar to the top. FALSE by default.}

\item{left_ui}{Custom left Ui content. Any element like dropdownMenu.}

\item{right_ui}{Custom right Ui content. Any element like dropdownMenu.}
}
\description{
Builds application header with given configurations and elements. It is called within "ui_header.R".
Check example application for detailed example
}
