% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{add_ui_right_sidebar}
\alias{add_ui_right_sidebar}
\title{add_ui_right_sidebar}
\usage{
add_ui_right_sidebar(
  sidebar_elements = NULL,
  collapsed = TRUE,
  overlay = TRUE,
  skin = "light",
  pinned = FALSE,
  controlbar_menu = NULL
)
}
\arguments{
\item{sidebar_elements}{- List of shiny ui elements}

\item{collapsed}{- If TRUE, the sidebar will be collapsed on app startup.}

\item{overlay}{- Whether the sidebar covers the content when expanded. Default to TRUE}

\item{skin}{- Sidebar skin. "dark" or "light".}

\item{pinned}{- Whether to block the controlbar state (TRUE or FALSE). Default to NULL}

\item{controlbar_menu}{- Right sidebar elements}
}
\description{
Builds application right sidebar with given configurations and elements. It is called within "ui_right_sidebar.R".
Check example application for detailed example
}
\section{Shiny Usage}{

Call this function from \code{program/ui_right_sidebar.R} to set right sidebar parameters
}

\examples{
  library(shiny)
  library(bs4Dash)

  #Inside ui_right_sidebar.R
  sidebar_elements <- list(div(checkboxInput("checkMe", "Example Check")))
  controlbar_menu  <- controlbarMenu(id = "controlbarmenu",
                                     controlbarItem("Item 2", "Simple text"))
 # -- Register Right Sidebar Elements in the ORDER SHOWN in the UI
  add_ui_right_sidebar(sidebar_elements = sidebar_elements,
                       controlbar_menu  = controlbar_menu)

}
