% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announcementConfigurationsAddin.R
\name{announcementConfigurationsAddin}
\alias{announcementConfigurationsAddin}
\title{Build Announcement Module Configuration YAML File}
\usage{
announcementConfigurationsAddin()
}
\description{
Call this as an addin to build valid yaml file that is needed for running
announcements file
}
