% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appReset.R
\name{appReset}
\alias{appReset}
\title{appReset Module Server Function}
\usage{
appReset(id, reset_wait = 5000, alert_location = "bodyAlert", logger)
}
\arguments{
\item{id}{- the ID of the Module's UI element}

\item{reset_wait}{- period to wait before session reload (in milliseconds)}

\item{alert_location}{- where to display module related messages
(default = "bodyAlert")}

\item{logger}{- logger to use}
}
\description{
Server-side function for the appResetButton  This is a custom
high-functionality button for session reload. The server function is used to
 provide module configurations.
}
\section{Shiny Usage}{

This function is not called directly by consumers - it is accessed in
server_local.R (or similar file) using the same id provided in \code{appResetButton}:

\strong{\code{appReset(id = "appResetId", logger = ss_userAction.Log)}}
}

\examples{
# Inside server_local.R

#appReset(id = "appResetId", logger = ss_userAction.Log)

}
