% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_helpers_external.R
\name{create_application_dashboard}
\alias{create_application_dashboard}
\title{create_application_dashboard}
\usage{
create_application_dashboard()
}
\description{
Creates application final dashboard from application different settings and configurations.
It is called once in the application in "ui.R". It should not be modified or updated by user
}
