% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{create_right_sidebar}
\alias{create_right_sidebar}
\title{Add the right sidebar to an existing application}
\usage{
create_right_sidebar(location)
}
\arguments{
\item{location}{path of the existing periscope2 application.}
}
\value{
no return value, creates right sidebar related UI R file and updates related source call in ui.R
}
\description{
User can update an existing application that does not have a right side bar and add a new empty one using this function.
}
\details{
If conversion is successful, the following message will be returned
\emph{"Add right sidebar conversion was successful. File(s) updated: ui.R, ui_right_sidebar.R"}

If the function called on an application with an existing right bar, message
\emph{"Right sidebar already available, no conversion needed"}
will be returned with no conversion

If the passed location is invalid, empty, not exist or not a valid periscope2 application, nothing will be added
and a related error message will be printed in console
}
\seealso{
\link[periscope2]{create_left_sidebar}
}
