% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{createAlert}
\alias{createAlert}
\title{createAlert}
\usage{
createAlert(
  id = NULL,
  selector = NULL,
  options,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{- Anchor id}

\item{selector}{- jQuery selector. Allow more customization for the anchor (nested tags).}

\item{options}{- List of options to pass to the alert}

\item{session}{- Shiny session object.}
}
\description{
Created an alert panel in server to be displayed the given ui anchor
}
\examples{
         createAlert(id       = "sidebarRightAlert",
                     options  = list(title    = "Right Side",
                                     status   = "success",
                                     closable = TRUE,
                                     content  = "Example Basic Sidebar Alert"))

}
