% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_template.R
\name{.copy_fw_files}
\alias{.copy_fw_files}
\title{.copy_fw_files
  Create new application standard files}
\usage{
.copy_fw_files(
  newloc,
  usersep,
  left_sidebar = TRUE,
  right_sidebar = FALSE,
  footer = FALSE,
  sample_app = FALSE
)
}
\arguments{
\item{newloc}{- string represents new application location}

\item{usersep}{- string represents path separator based on running OS}

\item{left_sidebar}{- boolean to control copying left ui sidebar (default = TRUE)}

\item{right_sidebar}{- boolean to control copying right ui sidebar (default = FALSE)}

\item{footer}{- boolean to control copying footer ui (default = FALSE)}

\item{sample_app}{- boolean to control copying sample app files (default = FALSE)}
}
\value{
nothing
}
\description{
.copy_fw_files
  Create new application standard files
}
