% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_template.R
\name{.copy_program_files}
\alias{.copy_program_files}
\title{.copy_program_files
  Copy program folder files}
\usage{
.copy_program_files(
  newloc,
  usersep,
  sample_app,
  left_sidebar = TRUE,
  right_sidebar = FALSE,
  footer = FALSE
)
}
\arguments{
\item{newloc}{- string represents new application location}

\item{usersep}{- string represents path separator based on running OS}

\item{sample_app}{- boolean to control copying sample app files (default = FALSE)}

\item{left_sidebar}{- boolean to control copying left ui sidebar (default = TRUE)}

\item{right_sidebar}{- boolean to control copying right ui sidebar (default = FALSE)}

\item{footer}{- boolean to control copying footer ui (default = FALSE)}
}
\value{
nothing
}
\description{
.copy_program_files
  Copy program folder files
}
