% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_template.R
\name{.create_dirs}
\alias{.create_dirs}
\title{.create_dirs
  Create new application directories}
\usage{
.create_dirs(newloc, usersep)
}
\arguments{
\item{newloc}{- string represents new application location}

\item{usersep}{- string represents path separator based on running OS}
}
\value{
nothing
}
\description{
.create_dirs
  Create new application directories
}
