% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadableReactTable.R
\name{downloadableReactTable}
\alias{downloadableReactTable}
\title{downloadableReactTable module server function}
\usage{
downloadableReactTable(id, table_data)
}
\arguments{
\item{id}{the ID of the Module's UI element}

\item{table_data}{reactive expression (or parameter-less function) that acts as table data source}
}
\value{
Rendered react table
}
\description{
Server-side function for the downloadableReactTableUI.
}
\section{Shiny Usage}{

This function is not called directly by consumers - it is accessed in
server.R using the same id provided in \code{downloadableReactTableUI}:

\strong{\code{downloadableReactTable(id)}}
}

\examples{
if (interactive()) {
 library(shiny)
 library(periscope2)
 shinyApp(ui = fluidPage(fluidRow(column(width = 12,
   downloadableReactTableUI("object_id1",
                            downloadtypes = c("csv", "tsv"),
                            hovertext     = "Download the data here!",
                            contentHeight = "300px",
                            singleSelect  = TRUE)))),
   server = function(input, output) {
       downloadableReactTable(id         = "object_id1",
                              table_data = reactiveVal(mtcars))})
}

}
\seealso{
\link[periscope2]{downloadableReactTableUI}

\link[periscope2]{downloadFileButton}

\link[periscope2]{logViewerOutput}

\link[periscope2]{downloadFile}

\link[periscope2]{downloadFile_ValidateTypes}

\link[periscope2]{downloadFile_AvailableTypes}

\link[periscope2]{downloadablePlot}
}
