% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logViewer.R
\name{logViewer}
\alias{logViewer}
\title{logViewer Module Server Function}
\usage{
logViewer(id, logger)
}
\arguments{
\item{id}{- the ID of the Module's UI element}

\item{logger}{- action logs to be displayed}
}
\value{
Shiny table render expression containing the currently logged lines
}
\description{
Server-side function for the logViewerOutput  This is box with table displaying application logs.
The server function is used to provide module configurations.
}
\section{Shiny Usage}{

This function is not called directly by consumers - it is accessed in
server_local.R (or similar file) using the same id provided in \code{logViewerOutput}:

\strong{\code{logViewer(id = "logViewerId", logger = ss_userAction.Log)}}
}

\examples{
# Inside server_local.R

#logViewer(id = "logViewerId", logger = ss_userAction.Log)

}
\seealso{
\link[periscope2]{logViewerOutput}

\link[periscope2]{downloadFile}

\link[periscope2]{downloadFile_ValidateTypes}

\link[periscope2]{downloadFile_AvailableTypes}

\link[periscope2]{downloadablePlot}

\link[periscope2]{downloadFileButton}

\link[periscope2]{downloadableTable}
}
