% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logViewer.R
\name{logViewerOutput}
\alias{logViewerOutput}
\title{logViewerOutput}
\usage{
logViewerOutput(id)
}
\arguments{
\item{id}{character id for the object}
}
\description{
Creates a box with table containing logged user actions. Table contents are auto updated whenever a user action is
logged.
}
\section{Table columns}{

\itemize{
    \item action - the action that id logged in any place in app
    \item time   - action time
}
}

\section{Example}{

\code{logViewerOutput('logViewerId')}
}

\section{Shiny Usage}{

Add the log viewer box to your box list

It is paired with a call to \code{logViewer(id, logger)}
in server
}

\examples{
# Inside ui_body add the log viewer box to your box list

logViewerOutput('logViewerId')


}
