% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeConfigurationsAddin.R
\name{themeConfigurationsAddin}
\alias{themeConfigurationsAddin}
\title{Build application theme configuration YAML file}
\usage{
themeConfigurationsAddin()
}
\value{
launch gadget window
}
\description{
Call this as an addin to build valid yaml file that is needed for creating/updating application
periscope_style.yaml file. The generated file can be used in periscope2 app using \link[periscope2]{set_app_parameters}.
}
\details{
The method can be called directly via `R` console or via RStudio addins menu
}
\examples{
if (interactive()) {
   periscope2:::themeConfigurationsAddin()
}

}
\seealso{
\link[periscope2:set_app_parameters]{periscope2:set_app_parameters()}
}
