% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{add_ui_footer}
\alias{add_ui_footer}
\title{add_ui_footer}
\usage{
add_ui_footer(left, right, fixed)
}
\arguments{
\item{left}{- Left text}

\item{right}{- Right text}

\item{fixed}{- Whether to fix the navbar to the top. FALSE by default.}
}
\description{
Builds application footer with given configurations and elements. It is called within "ui_footer.R".
Check example application for detailed example
}
