% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{add_ui_left_sidebar}
\alias{add_ui_left_sidebar}
\title{add_ui_left_sidebar}
\usage{
add_ui_left_sidebar(
  sidebar_elements = NULL,
  skin = "light",
  status,
  elevation,
  collapsed,
  minified,
  expand_on_hover,
  fixed,
  sidebar_menu,
  custom_area
)
}
\arguments{
\item{sidebar_elements}{list of shiny ui elements}

\item{skin}{Sidebar skin. "dark" or "light".}

\item{status}{Sidebar status. Check \code{?bs4Dash::dashboardSidebar()} for list of valid values}

\item{elevation}{Sidebar elevation. 4 by default (until 5)}

\item{collapsed}{If TRUE, the sidebar will be collapsed on app startup.}

\item{minified}{Whether to slightly close the sidebar but still show item icons. Default to TRUE}

\item{expand_on_hover}{Whether to expand the sidebar om hover. TRUE by default.}

\item{fixed}{character Whether to fix the sidebar. Default to TRUE.}

\item{sidebar_menu}{sidebar menu items}

\item{custom_area}{Sidebar bottom space area. Only works if sidebar is fixed}
}
\description{
This function adds left sidebar configurations and elements. It is called within "ui_left_sidebar.R".
Check example application for detailed example
}
