% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announcementConfigurationsAddin.R
\name{announcementConfigurationsAddin}
\alias{announcementConfigurationsAddin}
\title{Build Announcement Module Configuration YAML File}
\usage{
announcementConfigurationsAddin()
}
\value{
launch gadget window
}
\description{
Call this as an addin to build valid yaml file that is needed for running
announcements module. The generated file can be used in periscope2 app using \link[periscope2]{set_app_parameters}.
}
\details{
The method can be called directly via \code{R} console or via RStudio addins menu
}
\examples{
if (interactive()) {
   periscope2:::announcementConfigurationsAddin()
}

}
\seealso{
\link[periscope2:set_app_parameters]{periscope2:set_app_parameters()}
}
