% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{create_left_sidebar}
\alias{create_left_sidebar}
\title{Add the left sidebar to an existing application}
\usage{
create_left_sidebar(location)
}
\arguments{
\item{location}{path of the existing periscope2 application.}
}
\description{
User can update an existing application that does not have a left side bar and add a new empty one using this function.
}
\details{
If the function called on an application with an existing left bar, message
\emph{"Left sidebar already available, no conversion needed"}
will be returned with no conversion

If the passed location is invalid, empty, not exist or not a valid periscope2 application, nothing will be added
and a related error message will be printed in console
}
\seealso{
\link[periscope2]{create_right_sidebar}
}
