% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{create_right_sidebar}
\alias{create_right_sidebar}
\title{Add the right sidebar to an existing application}
\usage{
create_right_sidebar(location)
}
\arguments{
\item{location}{path of the existing periscope2 application.}
}
\description{
If the passed location is invalid, empty, not exist or not a valid periscope2 application, nothing will be added
and a related error message will be printed in console
}
\seealso{
\link[periscope2]{create_left_sidebar}
}
