% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{createAlert}
\alias{createAlert}
\title{createAlert}
\usage{
createAlert(
  id = NULL,
  selector = NULL,
  options,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{- Anchor id}

\item{selector}{- jQuery selector. Allow more customization for the anchor (nested tags).}

\item{options}{- List of options to pass to the alert}

\item{session}{- Shiny session object.}
}
\description{
Created an alert panel in server to be displayed the given ui anchor
}
\section{Shiny Usage}{

Call this function from \code{program/server_local.R} or any other server file to setup needed alert
}

\examples{
  library(shiny)
  library(bs4Dash)
  # Inside server_local.R
  periscope2::createAlert(id       = "sidebarRightAlert",
                          options  = list(title    = "Right Side",
                                          status   = "success",
                                          closable = TRUE,
                                          content  = "Example Basic Sidebar Alert"))

}
\seealso{
\link[bs4Dash:createAlert]{bs4Dash:createAlert()}
}
