% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadableReactTable.R
\name{downloadableReactTable}
\alias{downloadableReactTable}
\title{downloadableReactTable module server function}
\usage{
downloadableReactTable(id, table_data, selection_mode = NULL)
}
\arguments{
\item{id}{the ID of the Module's UI element}

\item{table_data}{reactive expression (or parameter-less function) that acts as table data source}

\item{selection_mode}{to enable row selection, set \code{selection_mode} value to either "single" for single row
selection or "multiple" for multiple rows selection, case insensitive. Any other value will
disable row selection, (default = NULL). An additional column will be added to the table if
selection mode is enabled with rabio buttons for single row selection and checkboxes for
"multiple" rows selection mode.}
}
\value{
Rendered react table
}
\description{
Server-side function for the downloadableReactTableUI.
}
\section{Shiny Usage}{

This function is not called directly by consumers - it is accessed in
server.R using the same id provided in \code{downloadableReactTableUI}:

\strong{\code{downloadableReactTable(id)}}
}

\examples{
if (interactive()) {
 library(shiny)
 library(periscope2)
 shinyApp(ui = fluidPage(fluidRow(column(width = 12,
   downloadableReactTableUI("object_id1",
                            downloadtypes = c("csv", "tsv"),
                            hovertext     = "Download the data here!")))),
   server = function(input, output) {
       downloadableReactTable(id         = "object_id1",
                              table_data = reactiveVal(mtcars))})
}

}
\seealso{
\link[periscope2]{downloadableReactTableUI}

\link[periscope2]{downloadFileButton}

\link[periscope2]{logViewerOutput}

\link[periscope2]{downloadFile}

\link[periscope2]{downloadFile_ValidateTypes}

\link[periscope2]{downloadFile_AvailableTypes}

\link[periscope2]{downloadablePlot}
}
