% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{get_app_info}
\alias{get_app_info}
\title{get_app_info}
\usage{
get_app_info()
}
\value{
String of application info
}
\description{
Returns current application app info that is set in \code{set_app_parameters} method to
 be used in any place in the application
}
\section{Shiny Usage}{

Call this function from \code{program/server_local.R} or any other server file
}

\examples{
  library(shiny)
  library(periscope2)

  # Display application info
  observeEvent(input$app_info, {
              shinyalert(html                = TRUE,
                         showConfirmButton   = FALSE,
                         animation           = "slide-from-top",
                         closeOnClickOutside = TRUE,
                         text                = get_app_info(),
                         title               = get_app_title())
              })

}
\seealso{
\link[periscope2:set_app_parameters]{periscope2:set_app_parameters()}

\link[periscope2:add_ui_footer]{periscope2:add_ui_footer()}

\link[periscope2:add_ui_left_sidebar]{periscope2:add_ui_left_sidebar()}

\link[periscope2:add_ui_header]{periscope2:add_ui_header()}

\link[periscope2:add_ui_body]{periscope2:add_ui_body()}

\link[periscope2:add_ui_right_sidebar]{periscope2:add_ui_right_sidebar()}

\link[periscope2:ui_tooltip]{periscope2:ui_tooltip()}

\link[periscope2:get_app_info]{periscope2:get_app_info()}

\link[periscope2:get_app_title]{periscope2:get_app_title()}
}
