% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_helpers_external.R
\name{load_announcements}
\alias{load_announcements}
\title{load_announcements}
\usage{
load_announcements()
}
\value{
number of seconds an announcement should be staying in caller application
}
\description{
Reads and parses application announcements configurations in \code{config/announce.yaml}, then display announcements in
application header.
}
\details{
If announce.yaml does not exist or contains invalid configurations. Nothing will be displayed.
Closing announcements is caller application responsibility
}
\examples{
# announce_close_time <- load_announcements()
# if (!is.null(announce_close_time)) {
#     shinyjs::delay(announce_close_time,{bs4Dash::closeAlert("announceAlert")})
# }

}
