% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periscope.R
\docType{package}
\name{periscope2}
\alias{periscope2}
\title{Periscope2 Shiny Application Framework}
\description{
This package supports a ui-standardized environment as well as
a variety of convenience functions for shiny applications.  Base
reusable functionality as well as UI paradigms are included to ensure a
consistent user experience regardless of application or developer.
}
\details{
A gallery of example apps is hosted at \href{http://periscopeapps.org:3838}{http://periscopeapps.org}
}
\section{Function Overview}{


\emph{Create a new framework application instance:\cr}
\link[periscope2]{create_application}\cr

\emph{Set application parameters in program/global.R:\cr}
\link[periscope2]{set_app_parameters}\cr

\emph{Get any url parameters passed to the application:\cr}
\link[periscope2]{get_url_parameters}\cr

\emph{Register user-created UI objects to the requisite application locations:\cr}
\link[periscope2]{create_left_sidebar}\cr
\link[periscope2]{create_right_sidebar}\cr
\link[periscope2]{add_ui_header}\cr
\link[periscope2]{add_ui_body}\cr
\link[periscope2]{add_ui_footer}

\emph{Included shiny modules with a customized UI:\cr}
\link[periscope2]{downloadFileButton}\cr
\link[periscope2]{downloadableTableUI}\cr
\link[periscope2]{downloadablePlotUI}

\emph{High-functionality standardized tooltips:\cr}
\link[periscope2]{ui_tooltip}
}

\section{More Information}{

\code{browseVignettes(package = 'periscope2')}
}

