% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{ui_tooltip}
\alias{ui_tooltip}
\title{ui_tooltip}
\usage{
ui_tooltip(id, label = "", text = "", placement = "top")
}
\arguments{
\item{id}{- The id for the tooltip object}

\item{label}{- text label to appear to the left of the tooltip image}

\item{text}{- Tooltip text shown when the user hovers over the image}

\item{placement}{- Where to display tooltip label. Available places are "top", "bottom", "left", "right" (default is "top")}
}
\description{
This function inserts a standardized tooltip image, label (optional),
and hovertext into the application UI
}
